rpc: Fetch correct funding streams in `getblocksubsidy`
The previous logic was indexing into the global funding streams list
with the local index into the active funding streams. This worked fine
for the first funding stream period, but once a funding stream ended and
another was enabled, the indices no longer matched.

We now return the `FundingStream` from `GetActiveFundingStreams`, as the
only user of that method is `getblocksubsidy`.