getinfo rpc: add lightwalletddisabled boolean to result
Let `lightwalletd` easily discover if the `zcashd` it's connected to is
configured with the required `lightwalletd=1` configuration option. The
new `getinfo` rpc field added by this PR, `lightwalletddisabled`,
is "backward" (disabled instead of enabled) because if this field
isn't returned at all (the running `zcashd` doesn't have this
commit), `lightwalletd` will receive it as `false` (missing fields
get the "default" value), so it will consider everything to be okay
(`lightwalletd` will start up). This allows `zcashd` to be updated after
`lightwalletd`.

Note that `zebrad` always supports `lightwalletd` (no special
configuration option required), so no change is needed there, although,
for consistency, we may want to add this field there and have it always
return `false` (lightwalletd support is never disabled).