build: python 3 compatibility
Ubuntu 16.04 "xenial xerus" does not come with Python 2.x by default.
It is possible to install a python-2.7 package, but this has its own
problem: no `python` or `python2` symlink (see bitcoin/bitcoin#7717).

This fixes the following scripts to work with python 3:
- `make check` (bctest,py, bitcoin-util-test.py)
- `make translate` (extract_strings_qt.py)
- `make symbols-check` (symbol-check.py)
- `make security-check` (security-check.py)

Explicitly call the python commands using $(PYTHON) instead
of relying on the interpreter line at the top of the scripts.

Zcash:
- Omit changes for qt and macOS build.
- Correct a bug introduced in symbol-check.py
  (keys in MAX_VERSIONS need to be byte strings).

Author: Wladimir J. van der Laan <laanwj@gmail.com>
Signed-off-by: Daira Hopwood <daira@jacaranda.org>