Prefer to create v5 transactions over v4 if either is possible
The `-preferredtxversion` config option was added in zcash/zcash#5967 to
enable node users to control which transaction version was used for new
transactions when the user requirements on its structure were supported
by more than one format. We set the initial default (in 2022) to the v4
transaction format, to account for the wallet ecosystem potentially not
having pervasive v5 transaction support at that time. Over 3 years later
we can now assume v5 transaction support is pervasive, and thus can
update the default for the zcashd wallet.